# backend/docs/openapi_snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Housing Society Management API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/societies/signup": {
      "post": {
        "tags": [
          "societies"
        ],
        "summary": "Signup",
        "description": "Public \u2014 no auth required. Creates a PENDING society + PENDING Admin\ntogether; both wait for Platform Owner approval (Section 26).",
        "operationId": "signup_api_v1_societies_signup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SocietySignupIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SocietySignupOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/societies": {
      "get": {
        "tags": [
          "societies"
        ],
        "summary": "List All",
        "operationId": "list_all_api_v1_societies_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SocietyOut"
                  },
                  "type": "array",
                  "title": "Response List All Api V1 Societies Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/societies/{society_id}/approve": {
      "post": {
        "tags": [
          "societies"
        ],
        "summary": "Approve",
        "operationId": "approve_api_v1_societies__society_id__approve_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "society_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Society Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SocietyOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/societies/{society_id}/status": {
      "patch": {
        "tags": [
          "societies"
        ],
        "summary": "Update Status",
        "operationId": "update_status_api_v1_societies__society_id__status_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "society_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Society Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SocietyStatusUpdateIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SocietyOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/residents/signup": {
      "post": {
        "tags": [
          "residents"
        ],
        "summary": "Signup",
        "description": "Public \u2014 no auth required. Resident waits for Admin approval\n(Section 26).",
        "operationId": "signup_api_v1_residents_signup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResidentSignupIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResidentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/residents/pending": {
      "get": {
        "tags": [
          "residents"
        ],
        "summary": "List Pending",
        "operationId": "list_pending_api_v1_residents_pending_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ResidentOut"
                  },
                  "type": "array",
                  "title": "Response List Pending Api V1 Residents Pending Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/residents/{resident_id}/approval": {
      "post": {
        "tags": [
          "residents"
        ],
        "summary": "Decide Approval",
        "operationId": "decide_approval_api_v1_residents__resident_id__approval_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "resident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Resident Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResidentApprovalIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResidentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/residents/property-links": {
      "post": {
        "tags": [
          "residents"
        ],
        "summary": "Link Property",
        "operationId": "link_property_api_v1_residents_property_links_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PropertyResidentLinkIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyResidentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/residents/property-links/{link_id}": {
      "delete": {
        "tags": [
          "residents"
        ],
        "summary": "Unlink Property",
        "operationId": "unlink_property_api_v1_residents_property_links__link_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "link_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Link Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyResidentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/residents/by-property/{property_id}": {
      "get": {
        "tags": [
          "residents"
        ],
        "summary": "Get Property Residents",
        "operationId": "get_property_residents_api_v1_residents_by_property__property_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "property_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Property Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PropertyResidentOut"
                  },
                  "title": "Response Get Property Residents Api V1 Residents By Property  Property Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/residents/{resident_id}/properties": {
      "get": {
        "tags": [
          "residents"
        ],
        "summary": "Get Resident Properties",
        "operationId": "get_resident_properties_api_v1_residents__resident_id__properties_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "resident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Resident Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PropertyResidentOut"
                  },
                  "title": "Response Get Resident Properties Api V1 Residents  Resident Id  Properties Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PropertyResidentLinkIn": {
        "properties": {
          "property_id": {
            "type": "string",
            "format": "uuid",
            "title": "Property Id"
          },
          "resident_id": {
            "type": "string",
            "format": "uuid",
            "title": "Resident Id"
          },
          "relationship_type": {
            "$ref": "#/components/schemas/RelationshipType"
          },
          "start_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          }
        },
        "type": "object",
        "required": [
          "property_id",
          "resident_id",
          "relationship_type"
        ],
        "title": "PropertyResidentLinkIn",
        "description": "Links an already-approved Resident to a property as Owner or Tenant\n(Section 12: multiple owners, owner+tenant coexist, resident can have\nmultiple properties)."
      },
      "PropertyResidentOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "property_id": {
            "type": "string",
            "format": "uuid",
            "title": "Property Id"
          },
          "resident_id": {
            "type": "string",
            "format": "uuid",
            "title": "Resident Id"
          },
          "relationship_type": {
            "$ref": "#/components/schemas/RelationshipType"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "start_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          }
        },
        "type": "object",
        "required": [
          "id",
          "property_id",
          "resident_id",
          "relationship_type",
          "is_active",
          "start_date",
          "end_date"
        ],
        "title": "PropertyResidentOut"
      },
...
      "RelationshipType": {
        "type": "string",
        "enum": [
          "OWNER",
          "TENANT"
        ],
        "title": "RelationshipType"
      },
      "ResidentApprovalIn": {
        "properties": {
          "approve": {
            "type": "boolean",
            "title": "Approve"
          },
          "rejection_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rejection Reason"
          }
        },
        "type": "object",
        "required": [
          "approve"
        ],
        "title": "ResidentApprovalIn"
      },
      "ResidentOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "society_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Society Id"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "mobile": {
            "type": "string",
            "title": "Mobile"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "status": {
            "$ref": "#/components/schemas/UserStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "society_id",
          "full_name",
          "mobile",
          "email",
          "status",
          "created_at"
        ],
        "title": "ResidentOut"
      },
      "ResidentSignupIn": {
        "properties": {
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "mobile": {
            "type": "string",
            "title": "Mobile"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "society_id": {
            "type": "string",
            "format": "uuid",
            "title": "Society Id"
          }
        },
        "type": "object",
        "required": [
          "full_name",
          "mobile",
          "society_id"
        ],
        "title": "ResidentSignupIn",
        "description": "Resident self-signup within a society (Admin approves/rejects after)."
      },
...
      "SocietyOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "status": {
            "$ref": "#/components/schemas/SocietyStatus"
          },
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          },
          "pincode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pincode"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "code",
          "status",
          "address",
          "city",
          "state",
          "pincode",
          "created_at"
        ],
        "title": "SocietyOut"
      },
      "SocietySignupIn": {
        "properties": {
          "society_name": {
            "type": "string",
            "title": "Society Name"
          },
          "society_code": {
            "type": "string",
            "title": "Society Code"
          },
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          },
          "pincode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pincode"
          },
          "admin_full_name": {
            "type": "string",
            "title": "Admin Full Name"
          },
          "admin_mobile": {
            "type": "string",
            "title": "Admin Mobile"
          },
          "admin_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Admin Email"
          }
        },
        "type": "object",
        "required": [
          "society_name",
          "society_code",
          "admin_full_name",
          "admin_mobile"
        ],
        "title": "SocietySignupIn",
        "description": "Admin self-signup for a new society (Section 26: Admin approval flow).\nCreates the Society (PENDING) and the Admin user (PENDING) together \u2014\nboth wait for Platform Owner approval before either becomes usable."
      },
      "SocietySignupOut": {
        "properties": {
          "society_id": {
            "type": "string",
            "format": "uuid",
            "title": "Society Id"
          },
          "admin_user_id": {
            "type": "string",
            "format": "uuid",
            "title": "Admin User Id"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Submitted \u2014 awaiting Platform Owner approval"
          }
        },
        "type": "object",
        "required": [
          "society_id",
          "admin_user_id"
        ],
        "title": "SocietySignupOut"
      },
      "SocietyStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "ACTIVE",
          "SUSPENDED"
        ],
        "title": "SocietyStatus"
      },
      "SocietyStatusUpdateIn": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/SocietyStatus"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "SocietyStatusUpdateIn"
      },
...
      "UserStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "ACTIVE",
          "REJECTED",
          "INACTIVE"
        ],
        "title": "UserStatus"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}